Doggo death

+++ Assets/Scripts/Rutger/Health.cs
@@ -7,7 +7,7 @@ public class Health : MonoBehaviour
 {
     public int maxHealth;
 
-    private int value = 20;
+    private float value = 20;
 
     public Action deadCallbacks;
 
@@ -21,18 +21,18 @@ public class Health : MonoBehaviour
         return value <= 0;
     }
 
-    public void Damage(int value)
+    public void Damage(float value)
     {
         this.value -= value;
         if (dead()) deadCallbacks();
     }
 
-    public void Heal(int value)
+    public void Heal(float value)
     {
         this.value = Mathf.Min(this.value + value, maxHealth);
     }
 
-    public int GetValue()
+    public float GetValue()
     {
         return value;
     }
